+++ .github/workflows/test_api_and_frontend.yaml
@@ -10,9 +10,6 @@ jobs:
   test:
     runs-on: ubuntu-latest
     timeout-minutes: 90
-    env:
-      FLAKY_TEST_PATHS: "skyportal/tests/frontend/sources_and_followup_etc/test_followup_requests.py skyportal/tests/api/test_photometry.py::test_token_user_big_post"
-
     services:
       postgres:
         image: postgres
@@ -163,31 +160,26 @@ jobs:
           pip install pre-commit
           pre-commit run --from-ref ${{ github.event.pull_request.base.sha }} --to-ref ${{ github.event.pull_request.head.sha }}
 
-      - name: Parse flaky file paths
-        run: |
-          export separator=" --ignore="
-          export regex="$( printf "${separator}%s" "${($FLAKY_TEST_PATHS)[@]}" )"
-
       - name: Run front-end tests part 1
         if: ${{ matrix.test_subset == 'frontend_pt1' }}
         run: |
-          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless 'skyportal/tests/frontend/sources_and_followup_etc ${regex}'
+          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless 'skyportal/tests/frontend/sources_and_followup_etc'
 
       - name: Run front-end tests part 2
         if: ${{ matrix.test_subset == 'frontend_pt2' }}
         run: |
-          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless '--ignore=skyportal/tests/frontend/sources_and_followup_etc skyportal/tests/frontend ${regex}'
+          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless '--ignore=skyportal/tests/frontend/sources_and_followup_etc skyportal/tests/frontend'
 
       - name: Run API & utils tests
         if: ${{ matrix.test_subset == 'api_and_utils' }}
         run: |
-          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless 'skyportal/tests/api skyportal/tests/tools skyportal/tests/utils skyportal/tests/rate_limiting ${regex}'
+          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless 'skyportal/tests/api skyportal/tests/tools skyportal/tests/utils skyportal/tests/rate_limiting'
 
       - name: Run flaky tests
         if: ${{ matrix.test_subset == 'flaky' }}
         continue-on-error: true
         run: |
-          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless '${FLAKY_TEST_PATHS}'
+          PYTHONPATH=. python baselayer/tools/test_frontend.py --xml --headless 'skyportal/tests/flaky'
 
       - name: Upload logs
         uses: actions/upload-artifact@v2
